+++ src/packages/press-area/computed.js
@@ -1,6 +1,6 @@
 import { fetchData } from 'press-ui/common/utils/fetch-data';
 
-const AREA_DATA_URL = 'https://tip-components-1251917893.cos.ap-guangzhou.myqcloud.com/rb/front-open-config__match__default__area_data.json';
+const AREA_DATA_URL = 'https://tip-components-1251917893.file.myqcloud.com/rb/front-open-config__match__default__area_data.json';
 
 function displayColumns(columns, columnsNum) {
   return columns.slice(0, +columnsNum);